pyproject.toml
[project]
name = "govcon-ai-pipeline"
version = "1.0.0"
description = "Production-ready multi-agent system for federal government contracting proposals (SDVOSB/VOSB/SB)"
authors = [
    {name = "The Bronze Shield", email = "dev@bronzeshield.com"}
]
readme = "README.md"
requires-python = ">=3.9"
license = {text = "Proprietary"}
keywords = ["govcon", "proposal", "agents", "ai", "sdvosb", "vosb", "federal"]
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
]

dependencies = [
    "anthropic>=0.39.0",
    "openai>=1.54.0",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg2-binary>=2.9.9",
    "asyncpg>=0.29.0",
    "greenlet>=3.2.0",
    "redis>=5.0.0",
    "qdrant-client>=1.11.0",
    "minio>=7.2.0",
    "pydantic>=2.9.0",
    "pydantic-settings>=2.5.0",
    "email-validator>=2.0.0",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "python-multipart>=0.0.12",
    "python-jose[cryptography]>=3.3.0",
    "passlib[bcrypt]>=1.7.4",
    "httpx>=0.27.0",
    "aiofiles>=24.1.0",
    "python-docx>=1.1.0",
    "python-dotenv>=1.0.0",
    "openpyxl>=3.1.0",
    "pandas>=2.2.0",
    "beautifulsoup4>=4.12.0",
    "lxml>=5.3.0",
    "pypdf>=5.0.0",
    "python-dateutil>=2.9.0",
    "pytz>=2024.1",
    "tenacity>=9.0.0",
    "click>=8.1.0",
    "rich>=13.9.0",
    "structlog>=24.4.0",
    "sentry-sdk[fastapi]>=2.14.0",
    "psutil>=6.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "pytest-cov>=6.0.0",
    "pytest-mock>=3.14.0",
    "httpx>=0.27.0",
    "faker>=30.0.0",
    "freezegun>=1.5.0",
    "mypy>=1.11.0",
    "ruff>=0.6.0",
    "black>=24.8.0",
    "isort>=5.13.0",
]

docs = [
    "mkdocs>=1.6.0",
    "mkdocs-material>=9.5.0",
    "mkdocstrings[python]>=0.26.0",
]

all = [
    "govcon-ai-pipeline[dev,docs]",
]

[project.scripts]
govcon = "govcon.cli:main"

[project.urls]
Homepage = "https://github.com/the-bronze-shield/govcon-ai-pipeline"
Documentation = "https://github.com/the-bronze-shield/govcon-ai-pipeline/docs"
Repository = "https://github.com/the-bronze-shield/govcon-ai-pipeline"
Issues = "https://github.com/the-bronze-shield/govcon-ai-pipeline/issues"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/govcon"]

[tool.pytest.ini_options]
minversion = "8.0"
addopts = [
    "-ra",
    "-q",
    "--strict-markers",
    "--strict-config",
    "--cov=govcon",
    "--cov-report=term-missing",
    "--cov-report=html",
]
testpaths = ["tests"]
pythonpath = ["src"]
asyncio_mode = "auto"

[tool.coverage.run]
source = ["tests"]

[tool.coverage.report]
exclude_lines = [
    "pragma: no cover",
    "def __repr__",
    "raise AssertionError",
    "raise NotImplementedError",
    "if __name__ == .__main__.:",
    "if TYPE_CHECKING:",
    "@abstractmethod",
]

[tool.mypy]
python_version = "3.9"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
strict_equality = true
strict_concatenate = true

[[tool.mypy.overrides]]
module = [
    "minio.*",
    "qdrant_client.*",
    "docx.*",
    "openpyxl.*",
]
ignore_missing_imports = true

[tool.ruff]
target-version = "py39"
line-length = 100
indent-width = 4

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP",  # pyupgrade
    "ARG", # flake8-unused-arguments
    "SIM", # flake8-simplify
]
ignore = [
    "E501",  # line too long (handled by black)
    "B008",  # do not perform function calls in argument defaults
    "B905",  # zip strict parameter
]

[tool.ruff.lint.per-file-ignores]
"__init__.py" = ["F401"]
"tests/**/*.py" = ["ARG", "S101"]

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
skip-magic-trailing-comma = false
line-ending = "auto"

[tool.black]
line-length = 100
target-version = ['py39', 'py310', 'py311', 'py312']
include = '\.pyi?$'
extend-exclude = '''
/(
  \.eggs
  | \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
)/
'''

[tool.isort]
profile = "black"
line_length = 100
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
ensure_newline_before_comments = true
